Renamed module :core:util to :core:common
