+++ core/common/src/main/AndroidManifest.xml
