+++ core/common/src/main/java/de/julianschwers/diary/core/common/Instant.kt
@@ -1,4 +1,4 @@
-package de.julianschwers.diary.core.util
+package de.julianschwers.diary.core.common
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
