+++ core/common/src/main/java/de/julianschwers/diary/core/common/NavigationArgumentRegister.kt
@@ -1,4 +1,4 @@
-package de.julianschwers.diary.core.util
+package de.julianschwers.diary.core.common
 
 import kotlin.uuid.ExperimentalUuidApi
 import kotlin.uuid.Uuid
