+++ settings.gradle.kts
@@ -26,5 +26,5 @@ include(":core:model")
 include(":data:database")
 include(":data:repository")
 include(":feature:journal")
-include(":core:util")
+include(":core:common")
 include(":feature:timeline")
